Downloaded torrents are now checked to see if they're valid. Trackers in announce and announce-all are also checked to see if they are alive and have the torrent.
If the torrent can't be downloaded, a dead torrent is downloaded or the 
tracker can't be contacted, the torrent will be downloaded again two 
more times until it gives up and downloads a lower quality if available.

The default config file has been changed to ~/.pytvshows/config in 
preperation for a state file and to make the atomic saving neater. 
PyTVShows now only needs write access to the .pytvshows/ directory 
rather than the entire directory where the old pytvshows.cfg was under.

There is now much better config file error checking.